Remove space between card containers
The inline-block display for card containers are
rendered with a space in between, which is not
desirable because each card is 300px wide has a
margin of 20px. A body width of 960px should
suffice, but it does not because the third element
is pushed on a new line.

We removed any whitespace between the car divs
(containers) as a workaroud.